+++ src/main/java/tornadofx/Component.kt
@@ -420,22 +420,29 @@ abstract class UIComponent(viewTitle: String? = "", icon: Node? = null) : Compon
         whenRefreshed { onRefresh() }
     }
 
+    // If the UIComponent property is set, prefer this to the property. This makes it possible to do subdelegation
+    // using forwardWorkspaceActions inside other components like TabPane(https://github.com/edvin/tornadofx/issues/894)
+    internal val effectiveSavable: BooleanExpression get() = booleanBinding(savable, properties) { (properties["tornadofx.savable"] as? BooleanExpression)?.value ?: this.value }
+    internal val effectiveRefreshable: BooleanExpression get() = booleanBinding(refreshable, properties) { (properties["tornadofx.refreshable"] as? BooleanExpression)?.value ?: this.value }
+    internal val effectiveCreatable: BooleanExpression get() = booleanBinding(creatable, properties) { (properties["tornadofx.creatable"] as? BooleanExpression)?.value ?: this.value }
+    internal val effectiveDeletable: BooleanExpression get() = booleanBinding(deletable, properties) { (properties["tornadofx.deletable"] as? BooleanExpression)?.value ?: this.value }
+
     /**
      * Forward the Workspace button states and actions to the TabPane, which
      * in turn will forward these states and actions to whatever View is represented
      * by the currently active Tab.
      */
     fun StackPane.connectWorkspaceActions() {
-        savableWhen { savable }
+        savableWhen { effectiveSavable }
         whenSaved { onSave() }
 
-        creatableWhen { creatable }
+        creatableWhen { effectiveCreatable }
         whenCreated { onCreate() }
 
-        deletableWhen { deletable }
+        deletableWhen { effectiveDeletable }
         whenDeleted { onDelete() }
 
-        refreshableWhen { refreshable }
+        refreshableWhen { effectiveRefreshable }
         whenRefreshed { onRefresh() }
     }
 
@@ -447,16 +454,16 @@ abstract class UIComponent(viewTitle: String? = "", icon: Node? = null) : Compon
      * Workspace states and actions, hence voiding this call.
      */
     fun forwardWorkspaceActions(uiComponent: UIComponent) {
-        savableWhen { uiComponent.savable }
+        savableWhen { uiComponent.effectiveSavable }
         whenSaved { uiComponent.onSave() }
 
-        deletableWhen { uiComponent.deletable }
+        deletableWhen { uiComponent.effectiveDeletable }
         whenDeleted { uiComponent.onDelete() }
 
-        creatableWhen { uiComponent.creatable }
+        creatableWhen { uiComponent.effectiveCreatable }
         whenCreated { uiComponent.onCreate() }
 
-        refreshableWhen { uiComponent.refreshable }
+        refreshableWhen { uiComponent.effectiveRefreshable }
         whenRefreshed { uiComponent.onRefresh() }
     }