+++ CHANGELOG.md
@@ -10,8 +10,10 @@
 
 
 ### Changes
+
  - Opened up HttpClientEngine and HttpURLEngine for easier subclassing/configuration of the Rest engine
  - Deprecated `observableList()` in favour of `observableListOf()`, and `[List, Set, Map].observable()` in favour of `[List, Set, Map].asObservable()` to be consistent with the Kotlin standard library.
+ - Subdelegation of workspace button states (https://github.com/edvin/tornadofx/issues/894)
  
 ### Additions